requirements/base.txt
django==1.8.*

djangorestframework==3.3.3

# djangorestframework-jsonapi==2.0

# 12factor env
django-environ==0.4.0

# Time zones support
pytz==2016.4

# Background job support
celery==3.1.23

# Python-PostgreSQL Database Adapter
# This require a compilation against PostgreSQL librareis
# If you are using Windows for development, you may consider
# installing a pre-built binaries from here:
#   http://www.stickpeople.com/projects/python/win-psycopg/

psycopg2==2.6.1
